{"schema":"decodex/commit/1","summary":"Split add event validation modules","authority":"manual"}
